requirements.txt
opencv-python~=4.8.0.76
Pillow~=10.0.0
setuptools==68.1.2